requirements.txt
asgiref==3.5.2
astroid==2.11.6
autopep8==1.6.0
colorama==0.4.5
dill==0.3.5.1  
Django==4.1.1  
django-braces==1.15.0
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-filter==22.1
django-js-asset==2.0.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.8.0
pylint==2.14.4
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
tzdata==2022.2
wrapt==1.14.1